Added security

+++ src/main/java/com/Project/QuickHost/Service/RoomServiceImpl.java
@@ -3,12 +3,15 @@ package com.Project.QuickHost.Service;
 import com.Project.QuickHost.Dto.RoomDto;
 import com.Project.QuickHost.Entity.Hotel;
 import com.Project.QuickHost.Entity.Room;
+import com.Project.QuickHost.Entity.User;
 import com.Project.QuickHost.Repository.HotelRepo;
 import com.Project.QuickHost.Repository.RoomRepo;
+import com.Project.QuickHost.exception.UnAuthorisedException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,6 +32,13 @@ public class RoomServiceImpl implements RoomService {
     public RoomDto createRoom(Long Hotelid,RoomDto roomDto) {
         log.info("Creating new Room with Hotelid: {}",Hotelid);
         Hotel hotel=hotelRepo.findById(Hotelid).orElseThrow(()->new RuntimeException("Hotel not found iwht id {}"+Hotelid));
+
+        //if give info to owner of hotel only
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("Not the current owner of hotel with id "+hotel.getId());
+        }
+
         Room room=modelMapper.map(roomDto,Room.class);
         room.setHotel(hotel);
 
@@ -62,6 +72,11 @@ public class RoomServiceImpl implements RoomService {
     @Override
     public RoomDto updateRoomById(Long id, RoomDto room) {
         Room room1=roomRepo.findById(id).orElseThrow(()->new RuntimeException("Room not found with id: {}"+id));
+        //check if ur owner or not
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room1.getHotel().getOwner())){
+            throw new UnAuthorisedException("Not the current owner of hotel with id "+room1.getHotel().getId());
+        }
         log.info("udating room with id:{} ",id);
         modelMapper.map(room,room1);
         room1.setId(id);
@@ -73,6 +88,13 @@ public class RoomServiceImpl implements RoomService {
     @Transactional
     public void deleteRoomById(Long id) {
         Room room=roomRepo.findById(id).orElseThrow(()->new RuntimeException("Room not found with id: {}"+id));
+
+        //check if ur owner or not
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("Not the current owner of hotel with id "+room.getHotel().getId());
+        }
+
         log.info("deleting room with id:{}",id);
         inventoryService.deleteInventory(room);
